Echo Rewrite
Added an auth model and router that a dev can extend and use.

Added a dependency: Laravel SerializableClosure.

Now, the entire application can be serialized.

This allows for the app to be built and serialized. Then unpacked on each request and run. This is expected to be much quicker.

You actually do not need the dependency. But then, all handler methods must be declared within an EchoModel subclass.